third_party/WebKit/LayoutTests/security/cannot-read-self-from-file-expected.txt

CONSOLE ERROR: XMLHttpRequest cannot load cannot-read-self-from-file.html. Cross origin requests are only supported for protocol schemes: http, data, https.
 Documents loaded from file: shouldn't be able to access themselves via XHR.
This is a testharness.js-based test.
PASS file: should be a unique-origin protocol for XHR purposes 
Harness: the test ran to completion.
